Code cleanup, refactor, envvars
    - Frontend cleanup complete.
    - Backend cleanup complete
    - API endpoint env vars working
    - Sample spreadsheets work with pop_res.py
    - pop_res.py cleanup complete
    - Core features functioinal on localhost
