+++ backend/reservations/views/guests.py
@@ -1,0 +1,38 @@
+
+from rest_framework.decorators import api_view
+
+@api_view(['GET', 'POST'])
+def guest_list(request):
+    if request.method == 'GET':
+        data = Guest.objects.all()
+
+        serializer = GuestSerializer(data, context={'request': request}, many=True)
+
+        return Response(serializer.data)
+
+    elif request.method == 'POST':
+        serializer = GuestSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(status=status.HTTP_201_CREATED)
+
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+@api_view(['PUT', 'DELETE'])
+def guest_detail(request, pk):
+    try:
+        guest = Guest.objects.get(pk=pk)
+    except Guest.DoesNotExist:
+        return Response(status=status.HTTP_404_NOT_FOUND)
+
+    if request.method == 'PUT':
+        serializer = GuestSerializer(guest, data=request.data,context={'request': request})
+        if serializer.is_valid():
+            serializer.save()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    elif request.method == 'DELETE':
+        guest.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
+
