+++ frontend/src/App.js
@@ -1,5 +1,3 @@
-//import RoomDataTable from './components/dasDataTable.js';
-//import "./styles/RoomDataTable.css";
 import RoomSelekta from "./pages/RoomSelekta"
 import Login from "./pages/Login"
 import { Route, Routes } from "react-router-dom"
@@ -15,18 +13,5 @@ export default function App() {
       <Route path="/rooms" element={<RoomSelekta />} />
     </Routes>
 
-
-    // <span className="componentContainer">
-
-    //     <div className="AppHeader">
-    //       <p>
-    //          Select it. Swap it. Wtvr.
-    //       </p>
-    //     </div>
-    //      
-    //     <div className="DTApp">
-    //       <RoomDataTable/>
-    //     </div>
-    // </span>
   );
 }
