+++ samples/exampleMainGuestList.csv
@@ -1,0 +1,11 @@
+ticket_code,last_name,first_name,email,product,
+asdf123,cc00ff33,Alice,alice@noop.com,Standard Room (2 Queen Beds)
+zasv123,cc00ff77,Zebra,zebra@noop.com,Knew Management Suite (1 King Bed)
+qwer123,cc00ff44,Bob,bob@noop.com,Babyface Suite (1 King Bed)
+zxcv123,cc00ff55,Charlie,charlie@noop.com,Babyface Suite (2 Queen Beds)
+zkkv123,cc00ff66,Charlie,charlie@noop.com,Babyface Suite (2 Queen Beds)
+zknm123,cc00ff77,Chad,chad@noop.com,Babyface Suite (2 Queen Beds)
+cvnm123,cc00ff88,Index,tyler32bit@gmail.com,Lakeview Standard Room (2 Queen Beds)
+cqwm143,cc00ff88,Index,tyler32bit@gmail.com,Babyface Suite (1 King Bed)
+cqwas43,cc00ff88,Index,tyler32bit@gmail.com,Babyface Suite (1 King Bed)
+cq12k43,cc00ff88,Index,tyler32bit@gmail.com,Babyface Suite (1 King Bed)
